+++ v2/index.js
@@ -17,7 +17,9 @@ api.route("/unit_contacts")
   .get(require("./unit_contacts"));
 
 api.route("/units")
-  .get(require("./units"));
+  .get(function(req, res, next) {
+    require("./units")(req, res, next);
+  });
 
 api.route("/fossils")
   .get(require("./fossils"));